fixed up unit tests

+++ tests/unit/modelSpec.cc
@@ -27,8 +27,8 @@ public:
     SET_VARS({{"x", "scalar"}});
 
     SET_DERIVED_PARAMS({
-        {"expDecay", [](const auto &pars, double dt) { return std::exp(-dt / pars.at("tau")); }},
-        {"init", [](const auto &pars, double) { return (std::exp(1) / pars.at("tau")); }}});
+        {"expDecay", [](const auto &pars, double dt) { return std::exp(-dt / pars.at("tau").cast<double>()); }},
+        {"init", [](const auto &pars, double) { return (std::exp(1) / pars.at("tau").cast<double>()); }}});
 };
 IMPLEMENT_SNIPPET(AlphaCurr);
