+++ tests/unit/postsynapticModels.cc
@@ -20,8 +20,8 @@ public:
     SET_PARAM_NAMES({"tau"});
 
     SET_DERIVED_PARAMS({
-        {"expDecay", [](const auto &pars, double dt){ return std::exp(-dt / pars.at("tau")); }},
-        {"init", [](const auto &pars, double dt){ return (pars.at("tau") * (1.0 - std::exp(-dt / pars.at("tau")))) * (1.0 / dt); }}});
+        {"expDecay", [](const auto &pars, double dt){ return std::exp(-dt / pars.at("tau").cast<double>()); }},
+        {"init", [](const auto &pars, double dt){ return (pars.at("tau").cast<double>() * (1.0 - std::exp(-dt / pars.at("tau").cast<double>()))) * (1.0 / dt); }}});
 };
 //--------------------------------------------------------------------------
 // Tests
